// types/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type CharacterConnection = {
	__typename?: "CharacterConnection";
	edges?: Maybe<Array<Maybe<CharacterEdge>>>;
	nodes?: Maybe<Array<Maybe<Character>>>;
	/** The pagination information */
	pageInfo?: Maybe<PageInfo>;
};
// ...
export type Favourites = {
	__typename?: "Favourites";
	/** Favourite anime */
	anime?: Maybe<MediaConnection>;
	/** Favourite characters */
	characters?: Maybe<CharacterConnection>;
	/** Favourite manga */
	manga?: Maybe<MediaConnection>;
	/** Favourite staff */
	staff?: Maybe<StaffConnection>;
	/** Favourite studios */
	studios?: Maybe<StudioConnection>;
};
// ...
export type MediaConnection = {
	__typename?: "MediaConnection";
	edges?: Maybe<Array<Maybe<MediaEdge>>>;
	nodes?: Maybe<Array<Maybe<Media>>>;
	/** The pagination information */
	pageInfo?: Maybe<PageInfo>;
};
// ...
export enum MediaFormat {
	/** Professionally published manga with more than one chapter */
	Manga = "MANGA",
	/** Anime movies with a theatrical release */
	Movie = "MOVIE",
	/** Short anime released as a music video */
	Music = "MUSIC",
	/** Written books released as a series of light novels */
	Novel = "NOVEL",
	/** (Original Net Animation) Anime that have been originally released online or are only available through streaming services. */
	Ona = "ONA",
	/** Manga with just one chapter */
	OneShot = "ONE_SHOT",
	/** (Original Video Animation) Anime that have been released directly on DVD/Blu-ray without originally going through a theatrical release or television broadcast */
	Ova = "OVA",
	/** Special episodes that have been included in DVD/Blu-ray releases, picture dramas, pilots, etc */
	Special = "SPECIAL",
	/** Anime broadcast on television */
	Tv = "TV",
	/** Anime which are under 15 minutes in length and broadcast on television */
	TvShort = "TV_SHORT",
}
// ...
export enum MediaSeason {
	/** Months September to November */
	Fall = "FALL",
	/** Months March to May */
	Spring = "SPRING",
	/** Months June to August */
	Summer = "SUMMER",
	/** Months December to February */
	Winter = "WINTER",
}

/** Media sort enums */
export enum MediaSort {
	Chapters = "CHAPTERS",
	ChaptersDesc = "CHAPTERS_DESC",
	Duration = "DURATION",
	DurationDesc = "DURATION_DESC",
	EndDate = "END_DATE",
	EndDateDesc = "END_DATE_DESC",
	Episodes = "EPISODES",
	EpisodesDesc = "EPISODES_DESC",
	Favourites = "FAVOURITES",
	FavouritesDesc = "FAVOURITES_DESC",
	Format = "FORMAT",
	FormatDesc = "FORMAT_DESC",
	Id = "ID",
	IdDesc = "ID_DESC",
	Popularity = "POPULARITY",
	PopularityDesc = "POPULARITY_DESC",
	Score = "SCORE",
	ScoreDesc = "SCORE_DESC",
	SearchMatch = "SEARCH_MATCH",
	StartDate = "START_DATE",
	StartDateDesc = "START_DATE_DESC",
	Status = "STATUS",
	StatusDesc = "STATUS_DESC",
	TitleEnglish = "TITLE_ENGLISH",
	TitleEnglishDesc = "TITLE_ENGLISH_DESC",
	TitleNative = "TITLE_NATIVE",
	TitleNativeDesc = "TITLE_NATIVE_DESC",
	TitleRomaji = "TITLE_ROMAJI",
	TitleRomajiDesc = "TITLE_ROMAJI_DESC",
	Trending = "TRENDING",
	TrendingDesc = "TRENDING_DESC",
	Type = "TYPE",
	TypeDesc = "TYPE_DESC",
	UpdatedAt = "UPDATED_AT",
	UpdatedAtDesc = "UPDATED_AT_DESC",
	Volumes = "VOLUMES",
	VolumesDesc = "VOLUMES_DESC",
}
// ...
export enum MediaStatus {
	/** Ended before the work could be finished */
	Cancelled = "CANCELLED",
	/** Has completed and is no longer being released */
	Finished = "FINISHED",
	/** Version 2 only. Is currently paused from releasing and will resume at a later date */
	Hiatus = "HIATUS",
	/** To be released at a later date */
	NotYetReleased = "NOT_YET_RELEASED",
	/** Currently releasing */
	Releasing = "RELEASING",
}
// ...
export type StaffConnection = {
	__typename?: "StaffConnection";
	edges?: Maybe<Array<Maybe<StaffEdge>>>;
	nodes?: Maybe<Array<Maybe<Staff>>>;
	/** The pagination information */
	pageInfo?: Maybe<PageInfo>;
};
// ...
export type StudioConnection = {
	__typename?: "StudioConnection";
	edges?: Maybe<Array<Maybe<StudioEdge>>>;
	nodes?: Maybe<Array<Maybe<Studio>>>;
	/** The pagination information */
	pageInfo?: Maybe<PageInfo>;
};